hist and floatlist

+++ src/alignment/alignment-ros-1-tester.py
@@ -5,7 +5,7 @@ import cv2
 import alignment
 from sensor_msgs.msg import Image
 from cv_bridge import CvBridge
-from bearnav2.msg import Alignment, IntList
+from bearnav2.msg import Alignment, FloatList
 from dynamic_reconfigure.server import Server
 from bearnav2.cfg import AlignmentConfig
 
@@ -31,7 +31,7 @@ def callbackB(msg):
     m.uncertainty = uncertainty
     pub.publish(m)
 
-    hm = IntList()
+    hm = FloatList()
     hm.data = hist
     pub_hist.publish(hm)
 
@@ -41,7 +41,7 @@ if __name__ == "__main__":
     aligner = alignment.Alignment()
 
     pub = rospy.Publisher("alignment/output", Alignment, queue_size=0)
-    pub_hist = rospy.Publisher("histogram", IntList, queue_size=0)
+    pub_hist = rospy.Publisher("histogram", FloatList, queue_size=0)
 
     rospy.logwarn("subscibing")
     rospy.Subscriber("/camera_front/image_color", Image, callbackB, queue_size=1)
